[TASK] Split up "Help within TYPO3" chapter (#161)
Previously, the keyboard commands were already a separate page which was displayed as a link at the bottom of a page and can easily overlooked.
As the keyboard commands chapter is already a bit longer, the other topics have been moved to separate pages (instead of integrating the keyboard commands into the existing page).
This also leaves room for improvement of each topic.

Releases: main, 12.4